refactor: Move types
